feat: Loosens type restrictions
In particular, it removes strong type adherence to stated schema types. This allows returning any conforming type, which can help with mocking, etc. It also removes the Context typealias, in preference of an expected "Context" type & Scalar types, and removes the output/Scalar types from the TypeMap, as they are no longer needed.